requirements.txt
numpy==1.23.5
pandas==2.2.3
Pillow==11.0.0
streamlit==1.33.0
